Move files to IEnumerable folder.
